src/main/java/io/github/raeperd/realworld/domain/article/ArticleService.java
package io.github.raeperd.realworld.domain.article;

import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
public class ArticleService {

    private final ArticleRepository articleRepository;

    public ArticleService(ArticleRepository articleRepository) {
        this.articleRepository = articleRepository;
    }

    @Transactional
    public Article createArticle(Article article) {
        return articleRepository.save(article);
    }
}
